Started simplifying tracks
Tracks now don't store the spatial calibration.  Instead, this needs to be applied outside the track.

+++ src/main/java/wbif/sjx/common/Analysis/SpatialCalculators/EuclideanDistanceCalculator.java
@@ -11,9 +11,9 @@ public class EuclideanDistanceCalculator implements SpatialCalculator {
 
         dist.put(f[0],0d);
         for (int i=0;i<x.length;i++) {
-            double dx = x[i]-x[0];
-            double dy = y[i]-y[0];
-            double dz = z[i]-z[0];
+            double dx = (double) x[i]-(double) x[0];
+            double dy = (double) y[i]-(double) y[0];
+            double dz = (double) z[i]-(double) z[0];
 
             dist.put(f[i],Math.sqrt(dx*dx + dy*dy + dz*dz));
         }
